Add files via upload
These Files are the source code for Managing Patient Management System :
consists of Frontend Page Named as findex.php
The file named as fdetails.php is to enter the data through the form towards database.

NOTE: Just need to change the database name as per your's database name.

dis.php file is used to Make queris on to the database and retrieving the data on the next page after submitting the data through the form.

Database queries should match to your created database.
NOTE: Use Xampp server to connect the pages to your database. xampp allows you to create the database on its own myphpadmin page.